refactor(client-addsnapshotsform): reorganized snapshotform folder structure
