+++ client/src/components/CardAddSnapshotsForm/ScrapeResults/ScrapeResult/ScrapeResult.css
